🐛 fixed: approval reason function
