+++ README.md
@@ -1,5 +1,5 @@
 # Jud Sue Jud Jang API
 - [x] Feature 1 : Authentication
 - [x] Feature 2 : Requester
-- [ ] Feature 3 : Approve Requester
+- [x] Feature 3 : Approve Requester
 - [ ] Feature 4 : Purchaser
